feature/69 refactor: timeline 관련 컴포넌트 패키지 이동
