+++ front/src/router/index.js
@@ -2,7 +2,7 @@ import Vue from "vue";
 import VueRouter from "vue-router";
 import MapPage from "@/components/map/MapPage";
 import PostDetailPage from "@/components/post/PostDetailPage";
-import TimelinePage from "@/components/TimelinePage";
+import TimelinePage from "@/components/timeline/TimelinePage";
 
 Vue.use(VueRouter);
